Convert color to long.

+++ app/src/main/java/com/example/spritesheetskotlin/palette/DBColor.kt
@@ -1,3 +1,3 @@
 package com.example.spritesheetskotlin.palette
 
-data class DBColor(val id: Int, var name: String, var color: Int, var paletteId: Int)
+data class DBColor(val id: Int, var name: String, var color: Long, var paletteId: Int)
